Create helper scripts for development
The hack/build.sh script will build the operator and push it to the
registry.

The hack/init.sh script will deploy all of the initial resources,
including the operator itself.

The hack/cleanup.sh script will remove all of the resources created in
the hack/init.sh script as well as any resources created from the
defined CRDs.

Signed-off-by: Jared Hocutt <jaredhocutt@gmail.com>